requirement.txt
flask
twilio
gspread
google-auth
groq
python-dotenv
gunicorn